+++ packages/fcl-bundle/src/program.js
@@ -2,12 +2,16 @@ const {program} = require("commander")
 const buildModule = require("./build/build")
 const watchModule = require("./build/build-watch")
 const assert = require("assert")
+const getPackageJson = require("./get-package-json")
+const {resolve} = require("path")
 
 module.exports = packageConfig => package => {
+  let bundlerPackageJson = getPackageJson(resolve(__dirname, ".."))
+
   program
     .name("FCL Build Tool")
     .description("Zero-configuration CLI to build FCL packages")
-    .version("1.0.0")
+    .version(bundlerPackageJson.version)
     .option("-w, --watch", "Run the build in watch mode")
     .parse()