+++ packages/fcl/src/discovery/index.js
@@ -1,4 +1,5 @@
 import authn from "./services/authn"
+export {getDiscoveryService, makeDiscoveryServices} from "./utils"
 
 const discovery = {
   authn,